+++ CHANGELOG.rst
@@ -12,9 +12,11 @@ The format is based on `Keep a Changelog <https://keepachangelog.com>`_.
 
 Added
 ++++++
-- The ability to fit **multiple** latent phenotypes in the global epistasis models. This adds the ``n_latent_phenotypes`` flag to ``AbstractEpistasis`` models, and changes the calls to certain methods / properties of that abstract model class and its concrete subclasses.
+- The ability to fit **multiple** latent phenotypes in the global epistasis models. This adds the ``n_latent_phenotypes`` flag to ``AbstractEpistasis`` models, and changes calls to certain methods / properties of that abstract model class and its concrete subclasses.
 
 - Added ``AbstractEpistasis.aic`` property.
+
+- Added ``globalepistasis.fit_models``
 
 - Added ``MultiLatentSigmoidPhenotypeSimulator``.